Add CENNSO overlay
With this commit, the default UI is "branded" to CENNSO and it offers
the download of files up to 2Gb.

The introduction of the cennso-overlay directory should keep the impact
of rebasing to upstream changes to a minimum in the future. It is also
much easier to add follow diffs when the upstream changes affect only
very small files like "Dockerfile" (opposed to removing files from the
repository to sidestep "COPY ." directives in Dockerfile).